+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased Changes
 * Removed `type_layout` being exposed except for internal tests.
 * Fixed alignment offset not taking into account previously added alignment.
+* Added `std140::Writer`, for writing dynamically laid out types to buffers.
 
 ## 0.1.0 (2020-09-18)
 * Initial MVP release